requirements.txt
argparse
flask
pyjsend
pytest
pytest-flask
pyserial